# feat: enhance seeding reliability and performance with automatic fallback