--- a/.env.example
+++ b/.env.example
@@ -61,7 +61,7 @@ MAIL_HOST=127.0.0.1
 MAIL_PORT=2525
 MAIL_USERNAME=null
 MAIL_PASSWORD=null
-MAIL_FROM_ADDRESS="hello@example.com"
+MAIL_FROM_ADDRESS="user@example.com"
 MAIL_FROM_NAME="${APP_NAME}"
 
 AWS_ACCESS_KEY_ID=
@@ -86,3 +86,11 @@ AVAILABLE_IMAGES_DEFAULT_COPYRIGHT_URL="${APP_URL}"
 
 PICTURES_DISK=public
 PICTURES_DIRECTORY=pictures
+
+#   | Default behavior:
+#    | - Non-production environments: Use local images (true)
+#    | - Production environment: Use remote images with local fallback (false)
+#    |
+#    | Set FAKER_USE_LOCAL_IMAGES=false to force remote image downloads
+#    | Set FAKER_USE_LOCAL_IMAGES=true to force local image usage
+# FAKER_USE_LOCAL_IMAGES=false