--- a/api-client/package.json
+++ b/api-client/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@metanull/inventory-app-api-client",
-  "version": "1.1.19-dev.0715.1753",
+  "version": "1.1.21-dev+20250724.1804",
   "main": "index.js",
   "types": "index.d.ts",
   "description": "TypeScript-Axios client for inventory-app API",
@@ -11,4 +11,3 @@
   "author": "Pascal Havelange",
   "license": "MIT"
 }
-