--- a/bootstrap/providers.php
+++ b/bootstrap/providers.php
@@ -2,7 +2,7 @@
 
 return [
     App\Providers\AppServiceProvider::class,
-    App\Providers\LoremPicsumFakerServiceProvider::class,
+    App\Providers\LocalImageFakerServiceProvider::class,
     App\Providers\FortifyServiceProvider::class,
     App\Providers\JetstreamServiceProvider::class,
     Intervention\Image\ImageServiceProvider::class,