--- a/config/mail.php
+++ b/config/mail.php
@@ -109,7 +109,7 @@ return [
     */
 
     'from' => [
-        'address' => env('MAIL_FROM_ADDRESS', 'hello@example.com'),
+        'address' => env('MAIL_FROM_ADDRESS', 'user@example.com'),
         'name' => env('MAIL_FROM_NAME', 'Example'),
     ],