--- a/database/seeders/CollectionSeeder.php
+++ b/database/seeders/CollectionSeeder.php
@@ -93,7 +93,5 @@ class CollectionSeeder extends Seeder
                 $item->update(['collection_id' => $collection->id]);
             }
         }
-
-        $this->command->info('CollectionSeeder completed successfully');
     }
 }